+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.15.0] - 2020-07-03
+
+### Fixed
+
+- Added an audible countdown to the timeout dialog for screenreader users
+- Added language parameter to the hmrcTimeoutDialog component to allow welsh language versions to be supported
+
 ## [1.14.1] - 2020-06-30
 
 ### Fixed
@@ -19,7 +26,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Updated implementation of hmrcHeader to be consistent with govuk-frontend v3.7.0 (an empty navigation
 array now treated as the same as navigation not being specified at all)
-- Added language parameter to the hmrcTimeoutDialog component to allow welsh language versions to be supported
 
 ### Compatible with